Builder in Java v2
This chnage does a few things. Maybe too many things for one change.
1. I like the approach they took in refactoring.guru, where the Director class doesn't keep a builder (from its constructor) but instead has an empty constructor and takes a builder per bike build. This way we don't have to new up a bunch of directors if we want to build a bunch of different bikes. All the Director needs to worry about is building a bike.
1. One of the big things I missed in v1 was that the BikeBuilder shouldn't have a getBike() method on it (even though it *does* in the Go implementation). The concrete builders are the ones that need to know about the resulting type. It shouldn't have to be the same kind of Bike returned every time, just the same interface to build a bike. So this change removes that method from the interface and passes it down to FixieBikeBuilder and TouringBikeBuilder to then return a FixedGearBike or TouringBike, instead of a Bike.
1. Another thing I noticed, is that the example on refactoring.guru has
2 vectors, not one. In the context of my code, it would be like, the
Director could build a racingBike or touringBike and the builders would
be FixedGear or 10Speed or 21Speed or Bike Manual. And THEN it would make sense for the
builder methods to take arguments. Right now, there's only 1 vector (i
don't know if vector is the right word here, but you get it), so the
Director doesn't have to worry about what the wheel size is, the builder
takes care of that. But I guess that's not really the builder pattern,
I'm realizing as I'm nearly done writing this gargantuan commit message.
1. I also turned on formatting, so some files got formatted without changes. Sorry...