Add 'add' script for fun
This adds a script that solves a problem I didn't really have but...it's fine. It was fun to put together. It can be used to maintain consistency in the file structure of the repository. It's for adding a couple folders if whoever it is wants to add a new implementation of a pattern.