Add products for abstract factor in java
The first step in implementing an Abstract Factory is mapping out the abstract and concrete products. In this example, I'm using flatware as the family of products and have a spoon, fork, and knife. In Java, the abstract products are interfaces and they are implemented in Modern, Classic, and Ornate forms.