Implement abstract, concrete factories in java
Products are in, now it's on to the factory classes that will create the products. First, the abstract factory: this class provides an interface that all factories will implement so that the caller won't have to worry about a specific factory implementation. Then the concrete factories: these are implementations of the abstract factory that take care of specific details of each product family behind the scenes.