Creates FlatwareApp and Client in Java abstract factory impl
This change creates the FlatwareApp that takes a factory as an argument and then uses the interface to build a set of flatware. This also creates the client that reads STDIN and creates a FlatwareApp with a factory based on the user input. Finally, this change updates the top-level README with the new pattern and the pattern README with implementation details.