Ignore transpiled .js files in ts projects
Currently, git is not ignoring .js files generated by tsc. We don't want to commit generated files to git, so this change updates .gitignore to include those files. Also, from now on, ts is the official dir name of any typescript impls. If that changes, .gitignore will need to be updated accordingly.