Add new functionality to add script
Again, this script is a bit frivilous but it's fun to maintain and add to. This change adds a piece that auto-creates a readme for the new pattern/language combo and creates the src/main/java file structure if it's a java implementation. Also good for me for getting the bash right on the first try.